Add exec_call opcode
